Fixed Submit not clicking by learning to scroll the webpage
It happens because my home pc has a bigger res 1440p screen and work pc has a 768p res. So naturally the home PC has a bigger viewport than office pc which is why "Submit" button is clearly visible on it and office pc screen  needs scrolling to see the Submit button to actually perform .click() action. It's because Selenium interacts with websites just like humans so even if it KNOWS the HTML page, it still needs to see and wait for the elements to be visible or clickable etc. to work properly.