Learnt how to handle downloads and uploads
Downloads needs chrome options' prefs to be used and a file path where the download will go to. Also added a static timer to check if the download has finished or took too long. Uploads work by sending keys as the directory path to the <input type='file'> tag on the upload button.